refactor: move link alert text below all buttons
