+++ src/components/TeacherComponents/TeacherShareButton/TeacherShareButton.tsx
@@ -1,36 +1,20 @@
-import { useState } from "react";
 import {
   OakFlex,
-  OakHeading,
-  OakIcon,
   OakSecondaryButton,
   OakSmallSecondaryButton,
 } from "@oaknational/oak-components";
 
 export const TeacherShareButton = ({
   label,
-  shareUrl,
-  shareActivated,
   variant,
+  handleClick,
+  shareUrl,
 }: {
   label: string;
-  shareUrl: string | null;
-  shareActivated?: () => void;
   variant: "primary" | "secondary";
+  handleClick: () => void;
+  shareUrl: string | null;
 }) => {
-  const [linkCopied, setLinkCopied] = useState(false);
-
-  const handleClick = () => {
-    if (!shareUrl) return;
-    if (!linkCopied && shareActivated) {
-      shareActivated();
-    }
-
-    setLinkCopied(true);
-    // copy url to clipboard
-    navigator.clipboard.writeText(shareUrl);
-  };
-
   if (variant === "primary") {
     return (
       <OakFlex $flexDirection={"column"} $position={"relative"}>
@@ -44,23 +28,6 @@ export const TeacherShareButton = ({
         >
           {label}
         </OakSecondaryButton>
-        {linkCopied && (
-          <OakFlex
-            $alignItems={"center"}
-            $position={"absolute"}
-            $top={["all-spacing-10", "all-spacing-11"]}
-          >
-            <OakIcon iconName={"tick"} $colorFilter={"text-success"} />
-            <OakHeading
-              tag="h2"
-              $font={"heading-light-7"}
-              $color={"text-success"}
-              aria-live="polite"
-            >
-              Link copied to clipboard
-            </OakHeading>
-          </OakFlex>
-        )}
       </OakFlex>
     );
   }
@@ -79,23 +46,6 @@ export const TeacherShareButton = ({
       >
         {label}
       </OakSmallSecondaryButton>
-      {linkCopied && (
-        <OakFlex
-          $alignItems={"center"}
-          $position={"absolute"}
-          $top={"all-spacing-8"}
-        >
-          <OakIcon iconName={"tick"} $colorFilter={"text-success"} />
-          <OakHeading
-            tag="h2"
-            $font={"heading-light-7"}
-            $color={"text-success"}
-            aria-live="polite"
-          >
-            Link copied to clipboard
-          </OakHeading>
-        </OakFlex>
-      )}
     </OakFlex>
   );
 };
